# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Documents Service API
  version: 0.1.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    UserIdHeader:
      type: apiKey
      in: header
      name: X-User-Id
...
security:
  - UserIdHeader: []
paths:
...
  /documents/upload-url:
    post:
      summary: Create document + version 1 and get a signed S3 upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folderId, title, filename, contentType, sizeBytes]
              properties:
                folderId: { type: string, format: uuid }
                title: { type: string }
                description: { type: string }
                filename: { type: string }
                contentType: { type: string }
                sizeBytes: { type: integer }
      responses:
        '201': { description: Created }
  /documents/{documentId}/versions:
    get:
      summary: List document versions
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
  /documents/{documentId}/versions/upload-url:
    post:
      summary: Create new document version and get a signed S3 upload URL
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Created }
  /documents/{documentId}/versions/{versionId}/download-url:
    get:
      summary: Get a signed S3 download URL
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: versionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
  /documents/{documentId}/notes:
    get:
      summary: List notes
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
    post:
      summary: Add a note
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Created }
...
  /documents/{documentId}/esign/docusign/envelopes:
    post:
      summary: Create e-sign envelope placeholder
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
...